15/07/2016/18/19

+++ app/src/main/java/ua/com/avatlantik/dubyk/i/dashboardclient/dto/SalesUGKTableDTO.java
@@ -6,19 +6,19 @@ package ua.com.avatlantik.dubyk.i.dashboardclient.dto;
 public class SalesUGKTableDTO {
 
     private String typeData;
-    private int sumMonth;
-    private int sumDay;
-    private int delta12;
-    private int delta3;
-    private int delta1;
+    private double sumMonth;
+    private double sumDay;
+    private double delta12;
+    private double delta3;
+    private double delta1;
 
     public SalesUGKTableDTO(String typeData, int sumMonth, int sumDay, int delta12, int delta3, int delta1) {
         this.typeData = typeData;
         this.sumMonth = sumMonth;
-        this.sumDay = sumDay;
-        this.delta12 = delta12;
-        this.delta3 = delta3;
-        this.delta1 = delta1;
+        this.sumDay   = sumDay;
+        this.delta12  = delta12;
+        this.delta3   = delta3;
+        this.delta1   = delta1;
     }
 
     public String getTypeData() {
@@ -29,43 +29,43 @@ public class SalesUGKTableDTO {
         this.typeData = typeData;
     }
 
-    public int getSumMonth() {
+    public double getSumMonth() {
         return sumMonth;
     }
 
-    public void setSumMonth(int sumMonth) {
+    public void setSumMonth(double sumMonth) {
         this.sumMonth = sumMonth;
     }
 
-    public int getDelta12() {
+    public double getDelta12() {
         return delta12;
     }
 
-    public void setDelta12(int delta12) {
+    public void setDelta12(double delta12) {
         this.delta12 = delta12;
     }
 
-    public int getSumDay() {
+    public double getSumDay() {
         return sumDay;
     }
 
-    public void setSumDay(int sumDay) {
+    public void setSumDay(double sumDay) {
         this.sumDay = sumDay;
     }
 
-    public int getDelta3() {
+    public double getDelta3() {
         return delta3;
     }
 
-    public void setDelta3(int delta3) {
+    public void setDelta3(double delta3) {
         this.delta3 = delta3;
     }
 
-    public int getDelta1() {
+    public double getDelta1() {
         return delta1;
     }
 
-    public void setDelta1(int delta1) {
+    public void setDelta1(double delta1) {
         this.delta1 = delta1;
     }
 }
